# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "QA Playground REST API",
    "version": "1.0.0",
    "description": "REST API designed for practicing manual, API, automation, and performance testing. Includes comprehensive error handling, JWT auth with refresh tokens, role-based access control, cart, products, and order management.",
    "contact": {
      "name": "QA Playground Team",
      "email": "support@qaplayground.local"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Local Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT access token obtained from /api/auth/login or /api/auth/register"
      }
    },
    "schemas": {
...
      "CartItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "productId": {
            "type": "integer",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Wireless Bluetooth Headphones"
          },
          "price": {
            "type": "number",
            "example": 79.99
          },
          "quantity": {
            "type": "integer",
            "example": 2
          },
          "lineTotal": {
            "type": "number",
            "example": 159.98
          },
          "stock": {
            "type": "integer",
            "example": 150
          },
          "category": {
            "type": "string",
            "example": "Electronics"
          },
          "imageUrl": {
            "type": "string",
            "example": "/images/products/p-001.jpg"
          }
        }
      },
      "Cart": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CartItem"
            }
          },
          "itemCount": {
            "type": "integer",
            "example": 2
          },
          "total": {
            "type": "number",
            "example": 159.98
          }
        }
      },
      "OrderItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "order_id": {
            "type": "integer",
            "example": 1
          },
          "product_id": {
            "type": "integer",
            "example": 1
          },
          "product_name": {
            "type": "string",
            "example": "Wireless Bluetooth Headphones"
          },
          "product_price": {
            "type": "number",
            "example": 79.99
          },
          "quantity": {
            "type": "integer",
            "example": 2
          },
          "imageUrl": {
            "type": "string",
            "example": "/images/products/p-001.jpg"
          }
        }
      },
      "Order": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "user_id": {
            "type": "integer",
            "example": 2
          },
          "total": {
            "type": "number",
            "example": 159.98
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "confirmed",
              "shipped",
              "delivered",
              "cancelled"
            ],
            "example": "pending"
          },
          "created_at": {
            "type": "string",
            "example": "2026-10-02 04:09:01"
          },
          "updated_at": {
            "type": "string",
            "example": "2026-10-02 04:09:01"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItem"
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/cart": {
      "get": {
        "summary": "Get current user cart",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Cart"
        ],
        "responses": {
          "200": {
            "description": "Cart object with items, line totals, and total amount",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "cart": {
                      "$ref": "#/components/schemas/Cart"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token"
          }
        }
      }
    },
    "/api/cart/items": {
      "post": {
        "summary": "Add item to current cart",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Cart"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "productId"
                ],
                "properties": {
                  "productId": {
                    "type": "integer",
                    "example": 1
                  },
                  "quantity": {
                    "type": "integer",
                    "example": 2,
                    "default": 1
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Item added to cart",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Item added to cart"
                    },
                    "item": {
                      "$ref": "#/components/schemas/CartItem"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Insufficient stock"
          },
          "401": {
            "description": "Missing or invalid token"
          },
          "404": {
            "description": "Product not found"
          },
          "422": {
            "description": "Validation failed"
          }
        }
      }
    },
    "/api/cart/items/{itemId}": {
      "delete": {
        "summary": "Remove item from cart by itemId",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Cart"
        ],
        "parameters": [
          {
            "name": "itemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Item removed from cart",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Item removed from cart"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token"
          },
          "404": {
            "description": "Cart or Cart item not found"
          }
        }
      }
    },
    "/api/orders": {
      "post": {
        "summary": "Create order from current user cart",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Orders"
        ],
        "responses": {
          "201": {
            "description": "Order placed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Order placed successfully"
                    },
                    "order": {
                      "$ref": "#/components/schemas/Order"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Cart is empty or insufficient product stock"
          },
          "401": {
            "description": "Missing or invalid token"
          }
        }
      },
      "get": {
        "summary": "Get orders (Own orders for regular users; all orders for admin)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Orders"
        ],
        "responses": {
          "200": {
            "description": "List of orders",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "orders": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Order"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token"
          }
        }
      }
    },
    "/api/orders/{id}": {
      "get": {
        "summary": "Get order details by order ID (Owner or Admin)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Orders"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Order details with line items",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "order": {
                      "$ref": "#/components/schemas/Order"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token"
          },
          "403": {
            "description": "Access denied"
          },
          "404": {
            "description": "Order not found"
          }
        }
      }
    },
    "/api/orders/{id}/cancel": {
      "patch": {
        "summary": "Cancel an order (Owner or Admin, only pending or confirmed)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Orders"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Order cancelled and stock restored",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Order cancelled successfully"
                    },
                    "order": {
                      "$ref": "#/components/schemas/Order"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Cannot cancel order with current status"
          },
          "401": {
            "description": "Missing or invalid token"
          },
          "403": {
            "description": "Access denied"
          },
          "404": {
            "description": "Order not found"
          }
        }
      }
    },
...
  }
}